Revert "Merge pull request #1 from stephen-cusi/NH"
This reverts commit 590f3308e369d610695cb9e1d1630a833d335f3d, reversing
changes made to e32d1771dc8070397e6b4d0d7afcb78fce8c4e10.

+++ game/client/particlemgr.cpp
@@ -1111,7 +1111,7 @@ bool CParticleMgr::Init(unsigned long count, IMaterialSystem *pMaterials)
 void CParticleMgr::Term()
 {
 	// Free all the effects.
-	intp iNext;
+	int iNext;
 	for ( intp i = m_Effects.Head(); i != m_Effects.InvalidIndex(); i = iNext )
 	{
 		iNext = m_Effects.Next( i );
@@ -1388,11 +1388,12 @@ void CParticleMgr::RemoveAllNewEffects()
 
 void CParticleMgr::RemoveAllEffects()
 {
-	intp iNext;
+	int iNext;
 	for ( intp i = m_Effects.Head(); i != m_Effects.InvalidIndex(); i = iNext )
 	{
 		iNext = m_Effects.Next( i );
 		RemoveEffect( m_Effects[i] );
+
 	}
 
 	RemoveAllNewEffects();
@@ -1969,7 +1970,7 @@ void CParticleMgr::UpdateAllEffects( float flTimeDelta )
 	m_bUpdatingEffects = false;
 
 	// Remove any effects that were flagged to be removed.
-	intp iNext;
+	int iNext;
 	for ( intp i = m_Effects.Head(); i != m_Effects.InvalidIndex(); i=iNext )
 	{
 		iNext = m_Effects.Next( i );
